Refactor client state to manage channels on server
