+++ include/IRC.hpp
@@ -18,6 +18,7 @@
 #include <stdlib.h>		// atoi
 #include <algorithm>
 #include <map>
+#include <set>
 
 #define RED "\033[1;31m" //-> for red color
 #define WHI "\033[0;37m" //-> for white color
